// src/codegen/ArenaDiscordIdentity.types.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@1.11.1.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export type Uint128 = string;
export interface InstantiateMsg {
	faucet_amount: Coin;
	owner: string;
}
export interface Coin {
	amount: Uint128;
	denom: string;
	[k: string]: unknown;
}
export type ExecuteMsg =
	| {
			set_profile: {
				addr: string;
				discord_profile: DiscordProfile;
			};
	  }
	| {
			set_connections: {
				connections: DiscordConnection[];
			};
	  }
	| {
			set_faucet_amount: {
				amount: Coin;
			};
	  }
	| {
			withdraw: {};
	  }
	| {
			update_ownership: Action;
	  };
export type Uint64 = string;
export type Action =
	| {
			transfer_ownership: {
				expiry?: Expiration | null;
				new_owner: string;
			};
	  }
	| "accept_ownership"
	| "renounce_ownership";
export type Expiration =
	| {
			at_height: number;
	  }
	| {
			at_time: Timestamp;
	  }
	| {
			never: {};
	  };
export type Timestamp = Uint64;
export interface DiscordProfile {
	avatar_hash?: string | null;
	user_id: Uint64;
	username: string;
}
export interface DiscordConnection {
	key: string;
	username: string;
}
export type QueryMsg =
	| {
			discord_profile: {
				addr: string;
			};
	  }
	| {
			connected_wallets: {
				discord_id: Uint64;
			};
	  }
	| {
			discord_connections: {
				addr: string;
			};
	  }
	| {
			ownership: {};
	  };
export type MigrateMsg = {
	from_compatible: {};
};
export type Addr = string;
export type ArrayOfAddr = Addr[];
export type ArrayOfDiscordConnection = DiscordConnection[];
export type NullableDiscordProfile = DiscordProfile | null;
export interface OwnershipForString {
	owner?: string | null;
	pending_expiry?: Expiration | null;
	pending_owner?: string | null;
}